// src-ts/modules/rs/network/_types.d.ts
import type { U64 } from "../../../utils/utils/_integerUtils";
export interface NetworkInterface {
    status: () => Promise<void>;
    ping: (url: string, timeout?: U64) => Promise<void>;
    resolve: (host: string) => Promise<void>;
    estimateBandwidth: (url?: string, sizeHintBytes?: U64, timeout?: U64) => Promise<void>;
    setMonitor: (interval: U64, targets?: string[]) => Promise<void>;
    stopMonitor: () => Promise<void>;
}
//# sourceMappingURL=_types.d.ts.map