// src-ts/modules/rs/notifications/_types.d.ts
export type NotificationPermission = "granted" | "denied" | "default" | string;
export interface NotificationsInterface {
    state: () => Promise<NotificationPermission>;
    request: () => Promise<Exclude<NotificationPermission, "default"> | string>;
    show: (title: string, body: string) => Promise<void>;
}
//# sourceMappingURL=_types.d.ts.map